+++ NEWs.md
@@ -1,4 +1,10 @@
 
+# metabolomicsR 1.1.0
+
+  - 2024-02-04: add new function 'QC_pipeline_df', and 'batch_norm_df', for a simplified data.frame with ID, batch, and metabolite columns.
+  - remove future::multiprocess, Warning message: object ‘multiprocess’ is not exported by 'namespace:future'
+
+
 
 # metabolomicsR 1.0.0